Restified the newsletter lambdas and cleaned code up
