+++ services/index.js
@@ -1,6 +1,7 @@
 import doPublish from './publish';
 import doContactUs from './contact_us';
-import doSignUp from './sign-up';
+import doSignUp from './mailchimp/sign-up/index';
+import doUpdateUser from './mailchimp/update-user/index';
 
 export function publish(event, context, cb) {
   try {
@@ -21,6 +22,14 @@ export function contactUs(event, context, cb) {
 export function signUp(event, context, cb) {
   try {
     doSignUp(event, cb);
+  } catch (e) {
+    cb(e);
+  }
+}
+
+export function updateUser(event, context, cb) {
+  try {
+    doUpdateUser(event, cb);
   } catch (e) {
     cb(e);
   }
