+++ services/serverless.yaml
@@ -40,8 +40,15 @@ functions:
     handler: index.signUp
     events:
       - http:
-          path: sign-up
+          path: mailing-list
           method: post
+          cors: true
+  updateUser:
+    handler: index.updateUser
+    events:
+      - http:
+          path: mailing-list
+          method: patch
           cors: true
